requirements.txt
Pillow>=9.0.0
svglib>=1.5.1
reportlab>=4.0.0
psutil>=7.0.0